PowerHAL: Cast the locktype value when calling new_lock_init
The function new_lock_init expects a lock type variable that
cannot be (by design) higher than the maximum value allowed
by a variable of type unsigned short.

In function halext_perf_lock_acquire, our lock_type variable
is an integer, hence needs casting to unsigned short when
passing it to the new_lock_init function.

+++ power/rqbalance_halext.c
@@ -441,7 +441,7 @@ int halext_perf_lock_acquire(struct rqbalance_halext_params *params)
 		lock_state = (params->argument[0] & 0xff);
 
 		return new_lock_init((unsigned int)params->time,
-					lock_type, lock_state);
+					(unsigned short)lock_type, lock_state);
 	}
 
 	for (i = 0; i <= arraysz; i++) {
